+++ model/labels/labels_dedupelabels_test.go
@@ -61,7 +61,7 @@ func TestVarint(t *testing.T) {
 func TestByteSize(t *testing.T) {
 	for _, testCase := range []struct {
 		lbls     Labels
-		expected int
+		expected uint64
 	}{
 		{
 			lbls:     FromStrings("__name__", "foo"),
